Improved client and offset height/width Added HTMLMeterElement stub Clean uni test

+++ LoboHTML/src/main/java/org/loboevolution/html/dom/domimpl/HTMLIFrameElementImpl.java
@@ -154,7 +154,7 @@ public class HTMLIFrameElementImpl extends HTMLElementImpl implements HTMLIFrame
 	}
 
 	@Override
-	public double getOffsetHeight() {
+	public Integer getOffsetHeight() {
 		return getClientHeight();
 	}
